Add shortcut to recover from broken LSP diagnostics
Sometimes the LSP server (or neovim client, no idea to be honest) gets
broken and the diagnostics are not refreshed anymore.

Lately I'm running into this problem very often while working with Go
and I find myself continuosly giving :LspStop followed by :LspStart (
because :LspRestart: restarts the server but doesn't removes the old
broken diagnostics).

This shortcut seems to be able to solve the issue (hope it will be fixed
upstream some day).